chore: altera estrututra e caminhos dos testes
